requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
fonttools==4.59.1
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.5
numpy==2.3.2
packaging==25.0
pillow==11.3.0
plyer==2.1.0
pyparsing==3.2.3
pystray==0.19.5
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
urllib3==2.5.0
